# mydocs-swagger/swagger.yaml
openapi: 3.0.0
info:
  title: My Swagger Learning
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 1.0.1
  contact:
    email: mohdayanansari199@gmail.com
    url: www.abc.com

# servers:
#     - url: http://api.example.com/v1
#       description: Optional server description, e.g. Main (production) server
#     - url: http://staging-api.example.com
#       description: Optional server description, e.g. Internal staging server for testing
servers:
  - url: "{protocol}://localhost:3050/api/{version}"
    description: Main (production) server
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http

  - url: http://staging-api.example.com
    description: Optional server description, e.g. Internal staging server for testing

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /ayan:
    get:
      tags:
        - String
      summary: returns a great message from ayan
      responses:
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: string
                example: "my string"
        400:
          description: Bad request
        500:
          description: Internal server error

  /object:
    get:
      tags:
        - Object
      summary: returns a unique course
      responses:
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /array:
    get:
      tags:
        - Array
      summary: returns all course
      responses:
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /course/{id}:
    get:
      tags:
        - String
      summary: returns course based on request id
      parameters:
        - name: id
          in: path
          required: true
          default: 22
          schema:
            type: string
      responses:
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /addCourse:
    post:
      tags:
        - Array
      summary: add a new course to existing courses
      consumes:
        - application/json
      produces:
        - application/json

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: boolean
        400:
          description: Bad request
        500:
          description: Internal server error
  /courseQuery:
    get:
      tags:
        - Array
      summary: trying to learn about query
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [delhi, london, jaipur]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]

      responses:
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string
        400:
          description: Bad request
        500:
          description: Internal server error
  /upload:
    post:
      tags:
        - Image
      summary: Uploading Course Image
      parameters: 
        - in: header
          name: auth
          
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: All good and success
        400:
          description: Bad request
        500:
          description: Internal server error